Reorganised tests and added @covers annotations
