+++ test/Reflection/ReflectionPropertyTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BetterReflectionTest;
+namespace BetterReflectionTest\Reflection;
 
 use BetterReflection\Reflector\ClassReflector;
 use BetterReflection\SourceLocator\ComposerSourceLocator;
@@ -18,6 +18,11 @@ class ReflectionPropertyTest extends \PHPUnit_Framework_TestCase
         $this->reflector = new ClassReflector(new ComposerSourceLocator($loader));
     }
 
+    /**
+     * @covers \BetterReflection\Reflection\ReflectionProperty::isPublic()
+     * @covers \BetterReflection\Reflection\ReflectionProperty::isProtected()
+     * @covers \BetterReflection\Reflection\ReflectionProperty::isPrivate()
+     */
     public function testVisibilityMethods()
     {
         $classInfo = $this->reflector->reflect('\BetterReflectionTest\Fixture\ExampleClass');
@@ -32,6 +37,9 @@ class ReflectionPropertyTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($publicProp->isPublic());
     }
 
+    /**
+     * @covers \BetterReflection\Reflection\ReflectionProperty::isStatic()
+     */
     public function testIsStatic()
     {
         $classInfo = $this->reflector->reflect('\BetterReflectionTest\Fixture\ExampleClass');
@@ -41,5 +49,29 @@ class ReflectionPropertyTest extends \PHPUnit_Framework_TestCase
 
         $staticProp = $classInfo->getProperty('publicStaticProperty');
         $this->assertTrue($staticProp->isStatic());
+    }
+
+    public function typesDataProvider()
+    {
+        return [
+            ['privateProperty', ['int', 'float', '\stdClass']],
+            ['protectedProperty', ['bool', 'bool[]', 'bool[][]']],
+            ['publicProperty', ['string']],
+        ];
+    }
+
+    /**
+     * @param string $propertyName
+     * @param string[] $expectedTypes
+     * @dataProvider typesDataProvider
+     * @covers \BetterReflection\Reflection\ReflectionProperty::getTypeStrings()
+     */
+    public function testGetDocBlockTypeStrings($propertyName, $expectedTypes)
+    {
+        $classInfo = $this->reflector->reflect('\BetterReflectionTest\Fixture\ExampleClass');
+
+        $property = $classInfo->getProperty($propertyName);
+
+        $this->assertSame($expectedTypes, $property->getDocBlockTypeStrings());
     }
 }
